Added additional variables,  loop and condition
The additional variables added are to find the tallest pokemon in the list. The additional loop added is to find the tallest pokemon and add the text "Wow, that's a big pokemon!" next to it if it is. The condition added is to make sure only one pokemon (the tallest) gets the message